refactor: autofill command
