+++ packages/sheets-ui/src/commands/commands/__tests__/auto-fill.controller.spec.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import type { ICellData, Injector, IStyleData, Nullable, Univer, Workbook } from '@univerjs/core';
 import {
     CellValueType,
     ICommandService,
@@ -27,6 +28,7 @@ import {
 import { DocSelectionManagerService } from '@univerjs/docs';
 import { EditorService, IEditorService } from '@univerjs/docs-ui';
 import { IRenderManagerService, RenderManagerService } from '@univerjs/engine-render';
+
 import {
     AddWorksheetMergeMutation,
     RemoveWorksheetMergeMutation,
@@ -34,10 +36,8 @@ import {
     SetSelectionsOperation,
     SheetsSelectionsService,
 } from '@univerjs/sheets';
-
 import { IPlatformService, IShortcutService, PlatformService, ShortcutService } from '@univerjs/ui';
 import { beforeEach, describe, expect, it } from 'vitest';
-import type { ICellData, Injector, IStyleData, Nullable, Univer, Workbook } from '@univerjs/core';
 import { AutoFillController } from '../../../controllers/auto-fill.controller';
 import { AutoFillService, IAutoFillService } from '../../../services/auto-fill/auto-fill.service';
 import { APPLY_TYPE } from '../../../services/auto-fill/type';
@@ -346,37 +346,38 @@ describe('Test auto fill rules in controller', () => {
                 const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
                 if (!workbook) throw new Error('This is an error');
                 // test number
-                (autoFillController as any)._triggerAutoFill(
-                    {
+
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 0,
                         startColumn: 0,
                         endRow: 0,
                         endColumn: 1,
                     },
-                    {
+                    targetRange: {
                         startRow: 0,
                         startColumn: 0,
                         endRow: 0,
                         endColumn: 3,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(0, 2)?.v).toBe(3);
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(0, 3)?.v).toBe(4);
                 // test number. descending
-                (autoFillController as any)._triggerAutoFill(
-                    {
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 1,
                         startColumn: 0,
                         endRow: 1,
                         endColumn: 1,
                     },
-                    {
+                    targetRange: {
                         startRow: 1,
                         startColumn: 0,
                         endRow: 1,
                         endColumn: 3,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(1, 2)?.v).toBe(0);
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(1, 3)?.v).toBe(-1);
             });
@@ -387,20 +388,20 @@ describe('Test auto fill rules in controller', () => {
                 const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
                 if (!workbook) throw new Error('This is an error');
                 // test extend number
-                (autoFillController as any)._triggerAutoFill(
-                    {
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 2,
                         startColumn: 0,
                         endRow: 2,
                         endColumn: 1,
                     },
-                    {
+                    targetRange: {
                         startRow: 2,
                         startColumn: 0,
                         endRow: 2,
                         endColumn: 3,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(2, 2)?.v).toBe('第 3');
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(2, 3)?.v).toBe('第 4');
             });
@@ -411,20 +412,20 @@ describe('Test auto fill rules in controller', () => {
                 const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
                 if (!workbook) throw new Error('This is an error');
                 // test chinese number
-                (autoFillController as any)._triggerAutoFill(
-                    {
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 3,
                         startColumn: 0,
                         endRow: 3,
                         endColumn: 2,
                     },
-                    {
+                    targetRange: {
                         startRow: 3,
                         startColumn: 0,
                         endRow: 3,
                         endColumn: 4,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(3, 3)?.v).toBe('五');
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(3, 4)?.v).toBe('日');
             });
@@ -436,20 +437,20 @@ describe('Test auto fill rules in controller', () => {
                 if (!workbook) throw new Error('This is an error');
 
                 // test chinese week
-                (autoFillController as any)._triggerAutoFill(
-                    {
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 4,
                         startColumn: 0,
                         endRow: 4,
                         endColumn: 2,
                     },
-                    {
+                    targetRange: {
                         startRow: 4,
                         startColumn: 0,
                         endRow: 4,
                         endColumn: 4,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(4, 3)?.v).toBe('星期五');
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(4, 4)?.v).toBe('星期日');
             });
@@ -460,20 +461,20 @@ describe('Test auto fill rules in controller', () => {
                 const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
                 if (!workbook) throw new Error('This is an error');
                 // test loop series
-                (autoFillController as any)._triggerAutoFill(
-                    {
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 5,
                         startColumn: 0,
                         endRow: 5,
                         endColumn: 1,
                     },
-                    {
+                    targetRange: {
                         startRow: 5,
                         startColumn: 0,
                         endRow: 5,
                         endColumn: 3,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(5, 2)?.v).toBe('丙');
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(5, 3)?.v).toBe('丁');
             });
@@ -484,20 +485,20 @@ describe('Test auto fill rules in controller', () => {
                 const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
                 if (!workbook) throw new Error('This is an error');
                 // test other string
-                (autoFillController as any)._triggerAutoFill(
-                    {
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 6,
                         startColumn: 0,
                         endRow: 6,
                         endColumn: 1,
                     },
-                    {
+                    targetRange: {
                         startRow: 6,
                         startColumn: 0,
                         endRow: 6,
                         endColumn: 3,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(6, 2)?.v).toBe('copy only');
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(6, 3)?.v).toBe('no series');
             });
@@ -508,20 +509,20 @@ describe('Test auto fill rules in controller', () => {
                 const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
                 if (!workbook) throw new Error('This is an error');
                 // test mixed mode
-                (autoFillController as any)._triggerAutoFill(
-                    {
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 7,
                         startColumn: 0,
                         endRow: 7,
                         endColumn: 3,
                     },
-                    {
+                    targetRange: {
                         startRow: 7,
                         startColumn: 0,
                         endRow: 7,
                         endColumn: 7,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(7, 4)?.v).toBe(3);
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(7, 5)?.v).toBe(4);
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(7, 6)?.v).toBe('第 3');
@@ -568,20 +569,20 @@ describe('Test auto fill rules in controller', () => {
                     },
                 }]);
 
-                (autoFillController as any)._triggerAutoFill(
-                    {
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 0,
                         startColumn: 5,
                         endRow: 2,
                         endColumn: 7,
                     },
-                    {
+                    targetRange: {
                         startRow: 0,
                         startColumn: 5,
                         endRow: 5,
                         endColumn: 7,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getActiveSheet()?.getMergeData().length).toBe(2);
                 expect(selectionManagerService.getCurrentLastSelection()?.primary).toEqual(expect.objectContaining({
                     startRow: 0,
@@ -619,20 +620,20 @@ describe('Test auto fill rules in controller', () => {
                     },
                 }]);
 
-                (autoFillController as any)._triggerAutoFill(
-                    {
+                commandService.executeCommand(AutoFillCommand.id, {
+                    sourceRange: {
                         startRow: 0,
                         startColumn: 0,
                         endRow: 5,
                         endColumn: 2,
                     },
-                    {
+                    targetRange: {
                         startRow: 0,
                         startColumn: 0,
                         endRow: 3,
                         endColumn: 2,
-                    }
-                );
+                    },
+                });
                 expect(workbook.getSheetBySheetId('sheet1')?.getCell(4, 0)?.v).toBeFalsy();
                 expect(selectionManagerService.getCurrentLastSelection()?.primary).toEqual(expect.objectContaining({
                     startRow: 0,
@@ -677,20 +678,20 @@ describe('Test auto fill rules in controller', () => {
             const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
             if (!workbook) throw new Error('This is an error');
             // test other string
-            (autoFillController as any)._triggerAutoFill(
-                {
+            commandService.executeCommand(AutoFillCommand.id, {
+                sourceRange: {
                     startRow: 14,
                     startColumn: 2,
                     endRow: 14,
                     endColumn: 4,
                 },
-                {
+                targetRange: {
                     startRow: 14,
                     startColumn: 0,
                     endRow: 14,
                     endColumn: 4,
-                }
-            );
+                },
+            });
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(14, 0)?.v).toBe(0.5);
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(14, 1)?.v).toBe(1);
             // undo redo
@@ -709,20 +710,20 @@ describe('Test auto fill rules in controller', () => {
             const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
             if (!workbook) throw new Error('This is an error');
             // equal ratio
-            (autoFillController as any)._triggerAutoFill(
-                {
+            commandService.executeCommand(AutoFillCommand.id, {
+                sourceRange: {
                     startRow: 14,
                     startColumn: 2,
                     endRow: 14,
                     endColumn: 4,
                 },
-                {
+                targetRange: {
                     startRow: 14,
                     startColumn: 2,
                     endRow: 14,
                     endColumn: 7,
-                }
-            );
+                },
+            });
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(14, 5)?.v).toBe(16);
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(14, 6)?.v).toBe(32);
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(14, 7)?.v).toBe(64);
@@ -757,20 +758,20 @@ describe('Test auto fill rules in controller', () => {
             const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
             if (!workbook) throw new Error('This is an error');
             // equal ratio
-            (autoFillController as any)._triggerAutoFill(
-                {
+            commandService.executeCommand(AutoFillCommand.id, {
+                sourceRange: {
                     startRow: 14,
                     startColumn: 2,
                     endRow: 14,
                     endColumn: 3,
                 },
-                {
+                targetRange: {
                     startRow: 14,
                     startColumn: 2,
                     endRow: 14,
                     endColumn: 4,
-                }
-            );
+                },
+            });
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(14, 4)?.v).toBe(6);
             const styles = getStyles(14, 4, 14, 4);
             expect(styles && styles[0][0]).toStrictEqual({
@@ -824,70 +825,70 @@ describe('Test auto fill rules in controller', () => {
             const workbook = get(IUniverInstanceService).getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET)!;
             if (!workbook) throw new Error('This is an error');
             // test right
-            (autoFillController as any)._triggerAutoFill(
-                {
+            commandService.executeCommand(AutoFillCommand.id, {
+                sourceRange: {
                     startRow: 16,
                     startColumn: 2,
                     endRow: 16,
                     endColumn: 2,
                 },
-                {
+                targetRange: {
                     startRow: 16,
                     startColumn: 2,
                     endRow: 16,
                     endColumn: 4,
-                }
-            );
+                },
+            });
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(16, 3)?.v).toBe(2);
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(16, 4)?.v).toBe(2);
             // test left
-            (autoFillController as any)._triggerAutoFill(
-                {
+            commandService.executeCommand(AutoFillCommand.id, {
+                sourceRange: {
                     startRow: 16,
                     startColumn: 2,
                     endRow: 16,
                     endColumn: 2,
                 },
-                {
+                targetRange: {
                     startRow: 16,
                     startColumn: 0,
                     endRow: 16,
                     endColumn: 2,
-                }
-            );
+                },
+            });
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(16, 1)?.v).toBe(2);
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(16, 0)?.v).toBe(2);
             // test up
-            (autoFillController as any)._triggerAutoFill(
-                {
+            commandService.executeCommand(AutoFillCommand.id, {
+                sourceRange: {
                     startRow: 16,
                     startColumn: 2,
                     endRow: 16,
                     endColumn: 2,
                 },
-                {
+                targetRange: {
                     startRow: 15,
                     startColumn: 2,
                     endRow: 16,
                     endColumn: 2,
-                }
-            );
+                },
+            });
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(15, 2)?.v).toBe(2);
             // test down
-            (autoFillController as any)._triggerAutoFill(
-                {
+            commandService.executeCommand(AutoFillCommand.id, {
+                sourceRange: {
                     startRow: 16,
                     startColumn: 2,
                     endRow: 16,
                     endColumn: 2,
                 },
-                {
+                targetRange: {
                     startRow: 16,
                     startColumn: 2,
                     endRow: 17,
                     endColumn: 2,
-                }
-            );
+                },
+            });
             expect(workbook.getSheetBySheetId('sheet1')?.getCell(17, 2)?.v).toBe(2);
         });
     });
