+++ packages/sheets-ui/src/commands/commands/refill.command.ts
@@ -15,10 +15,10 @@
  */
 
 import type { IAccessor, ICommand } from '@univerjs/core';
-import { CommandType } from '@univerjs/core';
-
-import { IAutoFillService } from '../../services/auto-fill/auto-fill.service';
 import type { APPLY_TYPE } from '../../services/auto-fill/type';
+
+import { CommandType } from '@univerjs/core';
+import { IAutoFillService } from '../../services/auto-fill/auto-fill.service';
 
 interface IRefillCommandParams {
     type: APPLY_TYPE;
@@ -29,7 +29,6 @@ export const RefillCommand: ICommand = {
     type: CommandType.COMMAND,
     handler: async (accessor: IAccessor, params: IRefillCommandParams) => {
         const autoFillService = accessor.get(IAutoFillService);
-        autoFillService.applyType = params.type;
-        return true;
+        return autoFillService.fillData(params.type);
     },
 };
