+++ packages/sheets-ui/src/controllers/auto-fill.controller.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import type { IAccessor, ICellData, ICommandInfo, IExecutionOptions, IMutationCommonParams, IMutationInfo, IRange, Nullable, UnitModel, Workbook } from '@univerjs/core';
+import type { ICellData, ICommandInfo, IExecutionOptions, IMutationCommonParams, IMutationInfo, IRange, Nullable, UnitModel, Workbook } from '@univerjs/core';
 import type {
     IAddWorksheetMergeMutationParams,
     IRemoveSheetMutationParams,
@@ -78,7 +78,7 @@ import { APPLY_TYPE, AutoFillHookType, DATA_TYPE } from '../services/auto-fill/t
 import { IEditorBridgeService } from '../services/editor-bridge.service';
 import { ISheetSelectionRenderService } from '../services/selection/base-selection-render.service';
 import { SheetsRenderService } from '../services/sheets-render.service';
-import { discreteRangeToRange, generateNullCellValueRowCol, rangeToDiscreteRange } from './utils/range-tools';
+import { discreteRangeToRange, generateNullCellValueRowCol } from './utils/range-tools';
 
 export class AutoFillController extends Disposable {
     private _beforeApplyData: Array<Array<Nullable<ICellData>>> = [];
@@ -113,7 +113,6 @@ export class AutoFillController extends Disposable {
     private _init() {
         this._initDefaultHook();
         this._onSelectionControlFillChanged();
-        this._onApplyTypeChanged();
         this._initQuitListener();
         this._initSkeletonChange();
     }
@@ -216,7 +215,7 @@ export class AutoFillController extends Disposable {
                                 endRow: filled.endRow,
                             };
 
-                            this._triggerAutoFill(source, selection);
+                            this._commandService.executeCommand(AutoFillCommand.id, { sourceRange: source, targetRange: selection });
                         })
                     )
                 );
@@ -265,135 +264,15 @@ export class AutoFillController extends Disposable {
         );
     }
 
-    // refill when apply type changed
-    private _onApplyTypeChanged() {
-        this.disposeWithMe(
-            toDisposable(
-                this._autoFillService.applyType$.subscribe(() => {
-                    this._handleFillData();
-                })
-            )
-        );
-    }
-
-    // eslint-disable-next-line max-lines-per-function, complexity
-    private _triggerAutoFill(source: IRange, selection: IRange) {
-        // if source range === dest range, do nothing;
-        if (
-            source.startColumn === selection.startColumn &&
-            source.startRow === selection.startRow &&
-            source.endColumn === selection.endColumn &&
-            source.endRow === selection.endRow
-        ) {
-            return;
-        }
-        // situation 1: drag to smaller range, horizontally.
-        if (selection.endColumn < source.endColumn && selection.endColumn > source.startColumn) {
-            this._commandService.executeCommand(AutoClearContentCommand.id, {
-                clearRange: {
-                    startRow: selection.startRow,
-                    endRow: selection.endRow,
-                    startColumn: selection.endColumn + 1,
-                    endColumn: source.endColumn,
-                },
-                selectionRange: selection,
-            });
-            return;
-        }
-        // situation 2: drag to smaller range, vertically.
-        if (selection.endRow < source.endRow && selection.endRow > source.startRow) {
-            this._commandService.executeCommand(AutoClearContentCommand.id, {
-                clearRange: {
-                    startRow: selection.endRow + 1,
-                    endRow: source.endRow,
-                    startColumn: selection.startColumn,
-                    endColumn: selection.endColumn,
-                },
-                selectionRange: selection,
-            });
-            return;
-        }
-        // situation 3: drag to larger range, expand to fill
-
-        // save ranges
-        const target = {
-            startRow: selection.startRow,
-            endRow: selection.endRow,
-            startColumn: selection.startColumn,
-            endColumn: selection.endColumn,
-        };
-        let direction: Nullable<Direction> = null;
-        if (selection.startRow < source.startRow) {
-            direction = Direction.UP;
-            target.endRow = source.startRow - 1;
-        } else if (selection.endRow > source.endRow) {
-            direction = Direction.DOWN;
-            target.startRow = source.endRow + 1;
-        } else if (selection.startColumn < source.startColumn) {
-            direction = Direction.LEFT;
-            target.endColumn = source.startColumn - 1;
-        } else if (selection.endColumn > source.endColumn) {
-            direction = Direction.RIGHT;
-            target.startColumn = source.endColumn + 1;
-        } else {
-            return;
-        }
-
-        const workbook = this._univerInstanceService.getCurrentUnitForType<Workbook>(UniverInstanceType.UNIVER_SHEET);
-        if (!workbook) return;
-
-        const unitId = workbook.getUnitId();
-        const subUnitId = workbook.getActiveSheet()?.getSheetId();
-        if (!subUnitId) return;
-
-        this._autoFillService.direction = direction;
-
-        const autoFillSource = this._injector.invoke((accessor: IAccessor) => rangeToDiscreteRange(source, accessor));
-        const autoFillTarget = this._injector.invoke((accessor: IAccessor) => rangeToDiscreteRange(target, accessor));
-
-        if (!autoFillSource || !autoFillTarget) {
-            return;
-        }
-        this._autoFillService.autoFillLocation = {
-            source: autoFillSource,
-            target: autoFillTarget,
-            unitId,
-            subUnitId,
-        };
-
-        const preferTypes: APPLY_TYPE[] = [];
-        const activeHooks = this._autoFillService.getActiveHooks();
-        activeHooks.forEach((hook) => {
-            const type = hook?.onBeforeFillData?.({ source: autoFillSource, target: autoFillTarget, unitId, subUnitId }, direction!);
-            if (type) {
-                preferTypes.unshift(type);
-            }
-        });
-
-        this._autoFillService.initPrevUndo();
-        // set apply type will trigger fillData
-        for (let i = 0; i < preferTypes.length; i++) {
-            const menuItem = this._autoFillService.menu.find((m) => m.value === preferTypes[i]);
-            if (menuItem && !menuItem.disable) {
-                this._autoFillService.applyType = preferTypes[i];
-                return;
-            }
-        }
-        // if no hook return apply type, use first available apply type
-        const first = this._autoFillService.menu.find((m) => m.disable === false)?.value;
-        if (first) {
-            this._autoFillService.applyType = first;
-        }
-    }
-
     private _handleDbClickFill(source: IRange) {
         const selection = this._detectFillRange(source);
         // double click only works when dest range is longer than source range
         if (selection.endRow <= source.endRow) {
             return;
         }
+
         // double click effect is the same as drag effect, but the apply area is automatically calculated (by method '_detectFillRange')
-        this._triggerAutoFill(source, selection);
+        this._commandService.executeCommand(AutoFillCommand.id, { sourceRange: source, targetRange: selection });
     }
 
     private _detectFillRange(source: IRange) {
@@ -436,13 +315,6 @@ export class AutoFillController extends Disposable {
             startRow,
             endRow: detectEndRow,
         };
-    }
-
-    private _handleFillData() {
-        if (!this._currentLocation) {
-            return;
-        }
-        this._commandService.executeCommand(AutoFillCommand.id, { unitId: this._currentLocation?.unitId, subUnitId: this._currentLocation?.subUnitId });
     }
 
     // calc apply data according to copy data and direction
